abstracted functions from Card modules to /src/modules
