+++ react_client/src/components/animation/Animation.jsx
@@ -1,0 +1,15 @@
+
+import ReactLoading from "react-loading";
+
+// Loading animation (GH)
+export const spinAnimation = function reactSpinLoadingAnimation() {
+  return (
+    <ReactLoading
+      className="mx-auto"
+      type="spin"
+      color="#4a4537"
+      height={100}
+      width={100}
+    />
+  );
+};
